\name{rLims-package}
\alias{lims-package}
\alias{rLims}
\docType{package}
\title{rLims}
\description{retrieve entry from www.mylims.org database}
\details{
\tabular{ll}{
Package: \tab lims\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-09\cr
License: \tab Unlimited \cr
}
Enjoy and share!
}
\author{Julien Wist <julien.wist@gmail.com>}
\references{Literature or other references for background information}

