\name{lims.getDate}
\alias{lims.getDate}
\title{
Set current date, user and time stamp for logging 
}
\description{
Create list object that contains "user" and "date". The date comes with a random time stamp to create unique entries.
}
\usage{
lims.getDate()
}
\details{
The date comes as "2014_02_23_63582836790" and can be used as a unique time stamp
}
\references{
www.mylims.org
}
\author{
Julien Wist
}

\examples{
lims.getDate()
}

\keyword{ lims }
\keyword{ nmr }
\keyword{ metabolomics }
\keyword{ log }
