\name{urlList}
\docType{data}
\alias{urlList}
\title{list of input URL to test rLims package}
\description{
A list of URL to test the package. Theses URL provides links to data stored in mylims database.
}
\usage{urlList}
\format{A list containing 5 URLs}
\source{mylims.univalle.edu.co}
\keyword{datasets}
